requirements_dev.txt
cmake
types-requests==2.28.0
types-PyYAML==6.0.11
pytest-cov==3.0.0
mypy==0.961
flake8==4.0.1
tox==3.25.1
pytest-xdist==2.5.0
